+++ mas-foundation/src/androidTest/java/com/ca/mas/foundation/MASFoundationTestSuite.java
@@ -14,6 +14,7 @@ import org.junit.runners.Suite;
 
 @RunWith(Categories.class)
 @Suite.SuiteClasses({
+        MASEnrollmentStartTest.class,
         MASAuthorizationCodeFlowTest.class,
         MASAuthorizationProviderTest.class,
         MASClaimsTest.class,
@@ -22,7 +23,6 @@ import org.junit.runners.Suite;
         MASDeviceIdentifierTest.class,
         MASDeviceTest.class,
         MASDynamicSDKTest.class,
-        MASEnrollmentStartTest.class,
         MASGeoFencingTest.class,
         MASJwtSigningTest.class,
         MASLoginTest.class,